# Enregister EH var that are single def