--- a/src/coreclr/jit/lclvars.cpp
+++ b/src/coreclr/jit/lclvars.cpp
@@ -2614,7 +2614,7 @@ void Compiler::lvaSetVarLiveInOutOfHandler(unsigned varNum)
         {
             noway_assert(lvaTable[i].lvIsStructField);
             lvaTable[i].lvLiveInOutOfHndlr = 1;
-            // For now, only enregister an EH Var if it is a single def.
+            // For now, only enregister an EH Var if it is a single def and whose refCnt > 1.
             if (!lvaEnregEHVars || !lvaTable[i].lvEhWriteThruCandidate || lvaTable[i].lvRefCnt() <= 1)
             {
                 lvaSetVarDoNotEnregister(i DEBUGARG(DNER_LiveInOutOfHandler));
@@ -2622,7 +2622,7 @@ void Compiler::lvaSetVarLiveInOutOfHandler(unsigned varNum)
         }
     }
 
-    // For now, only enregister an EH Var if it is a single def.
+    // For now, only enregister an EH Var if it is a single def and whose refCnt > 1.
     if (!lvaEnregEHVars || !varDsc->lvEhWriteThruCandidate || varDsc->lvRefCnt() <= 1)
     {
         lvaSetVarDoNotEnregister(varNum DEBUGARG(DNER_LiveInOutOfHandler));