--- a/src/coreclr/scripts/superpmi.py
+++ b/src/coreclr/scripts/superpmi.py
@@ -1565,7 +1565,7 @@ class SuperPMIReplayAsmDiffs:
             "COMPlus_JitDiffableDasm": "1",
             "COMPlus_JitEnableNoWayAssert": "1",
             "COMPlus_JitNoForceFallback": "1",
-            "COMPlus_JitDisasmWithGC": "1" }
+            "COMPlus_JitDisasmWithGC": "0" }
 
         if self.coreclr_args.gcinfo:
             asm_complus_vars.update({