--- a/src/libraries/System.Private.CoreLib/src/System/Threading/ExecutionContext.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Threading/ExecutionContext.cs
@@ -149,6 +149,12 @@ namespace System.Threading
 
         internal static void RunInternal(ExecutionContext? executionContext, ContextCallback callback, object? state)
         {
+            // Note: ExecutionContext.RunInternal is an extremely hot function and used by every await, ThreadPool execution, etc.
+            // Note: Manual enregistering may be addressed by "Exception Handling Write Through Optimization"
+            //       https://github.com/dotnet/runtime/blob/main/docs/design/features/eh-writethru.md
+
+            // Enregister previousExecutionCtx0 so they can be used in registers without EH forcing them to stack
+
             Thread currentThread = Thread.CurrentThread;
             ExecutionContext? previousExecutionCtx0 = currentThread._executionContext;
             if (previousExecutionCtx0 != null && previousExecutionCtx0.m_isDefault)