In https://github.com/dotnet/runtime/issues/35534, we saw mixed result of improvements and regressions. Hence the plan is to gradually enable this feature. This is the first step where we would enregister all EH vars in a method that are single def. In future, we will enable more scenarios.

Contributes to https://github.com/dotnet/runtime/issues/35923