@kunalspathak 2021-01-29
There was an issue in how we were setting `lvSingleDef` because of which several scenarios (including the one that @benaadams pointed) were not showing improvements. By fixing them, I was able to revert the work around that @benaadams  added without seeing regression.

### AsyncMethodBuilderCore.Start

#### Before:
|    Method |        Job |              Toolchain |       Mean |     Error |   StdDev |
|---------- |----------- |----------------------- |-----------:|----------:|---------:|
| Calculate | Job-CNAIDU |                CoreRun | 9,672.3 ns | 108.90 ns | 96.54 ns |

#### After:
|    Method |        Job |              Toolchain |       Mean |    Error |   StdDev |
|---------- |----------- |----------------------- |-----------:|---------:|---------:|
| Calculate | Job-NDWJCT |                CoreRun | 9,580.4 ns | 97.93 ns | 91.60 ns |

<details>
<summary>Calculate benchmark code</summary>

```c#
private const int OperationsPerInvoke = 10000;

[Benchmark(OperationsPerInvoke = OperationsPerInvoke)]
public async Task<int> Calculate()
{
    int result = 0;
    for (var i = 0; i < OperationsPerInvoke; i++)
    {
        result += await Task.Run(() => 1);
    }
    return result;
}
```
</details>

Asmdiff:  https://www.diffchecker.com/fOIS7SHG

### ExecutionContext.RunInternal

#### Before

| Method |        Job |              Toolchain |       Mean |     Error |    StdDev |
|------- |----------- |----------------------- |-----------:|----------:|----------:|
|    Run | Job-BRDNAB |                CoreRun | 662.891 ns | 5.8385 ns | 5.1756 ns |

#### After

| Method |        Job |              Toolchain |       Mean |     Error |    StdDev |
|------- |----------- |----------------------- |-----------:|----------:|----------:|
|    Run | Job-AKVWXZ |                CoreRun | 652.817 ns | 3.1342 ns | 2.9317 ns |


<details>
<summary>Calculate benchmark code</summary>

```c#
private const int OperationsPerInvoke = 1000;

[Benchmark(OperationsPerInvoke = OperationsPerInvoke)]
public void Run()
{
    var ec = ExecutionContext.Capture();

    for (var i = 0; i < OperationsPerInvoke; i++)
    {
        ExecutionContext.Run(ec, (o) => { }, null);
    }
}
```
</details>

Asmdiff: https://www.diffchecker.com/U8yRJWy8

FYI - @stephentoub