@kunalspathak 2021-03-08
I tried looking into some benchmarks that either has exception handling or use async/await and here are some of the results of % changes from this PR.

Benchmark: https://github.com/dotnet/performance/blob/4e14ecc54759224121b807c70e46fb1d9ee0e7d6/src/benchmarks/micro/libraries/System.Threading.Tasks/Perf.AsyncMethods.cs


Method | Before | After | Change
-- | -- | -- | --
EmptyAsyncMethodInvocation | 706.1 ns | 689.6 ns | -2.34%
SingleYieldMethodInvocation | 11,136.2 ns | 10,628.6 ns | -4.56%
Yield | 4,908.7 ns | 4,927.2 ns | 0.38%


Benchmark: https://github.com/dotnet/performance/blob/4e14ecc54759224121b807c70e46fb1d9ee0e7d6/src/benchmarks/micro/runtime/Exceptions/Handling.cs


Method | kind | Master | PR | Change
-- | -- | -- | -- | --
ThrowAndCatch | Software | 92.45 us | 92.49 us | 0.04%
ThrowAndCatch_ManyCatchBlocks | Software | 100.07 us | 100.80 us | 0.73%
ThrowAndCatchFinally | Software | 93.26 us | 93.32 us | 0.06%
ThrowAndCatchWhen | Software | 97.13 us | 95.47 us | -1.71%
ThrowAndCatchWhenFinally | Software | 96.65 us | 96.08 us | -0.59%
ThrowAndCatchDeep | Software | 270.15 us | 267.29 us | -1.06%
ThrowAndCatchDeepRecursive | Software | 291.16 us | 280.76 us | -3.57%
MultipleNestedTryCatch_FirstCatches | Software | 94.11 us | 91.68 us | -2.58%
MultipleNestedTryCatch_LastCatches | Software | 99.87 us | 98.23 us | -1.64%
MultipleNestedTryFinally | Software | 110.36 us | 108.39 us | -1.79%
CatchAndRethrowDeep | Software | 1,052.16 us | 1,046.31 us | -0.56%
CatchAndThrowOtherDeep | Software | 1,195.62 us | 1,188.44 us | -0.60%
TryAndFinallyDeep | Software | 332.00 us | 328.09 us | -1.18%
TryAndCatchDeep_CaugtAtTheTop | Software | 298.18 us | 291.30 us | -2.31%
ThrowAndCatch | Hardware | 101.62 us | 101.13 us | -0.48%
ThrowAndCatch_ManyCatchBlocks | Hardware | 110.26 us | 109.94 us | -0.29%
ThrowAndCatchFinally | Hardware | 102.48 us | 102.10 us | -0.37%
ThrowAndCatchWhen | Hardware | 106.19 us | 105.14 us | -0.99%
ThrowAndCatchWhenFinally | Hardware | 106.09 us | 106.79 us | 0.66%
ThrowAndCatchDeep | Hardware | 284.64 us | 279.61 us | -1.77%
ThrowAndCatchDeepRecursive | Hardware | 302.58 us | 295.29 us | -2.41%
MultipleNestedTryCatch_FirstCatches | Hardware | 102.21 us | 102.07 us | -0.14%
MultipleNestedTryCatch_LastCatches | Hardware | 107.70 us | 107.19 us | -0.47%
MultipleNestedTryFinally | Hardware | 118.17 us | 117.21 us | -0.81%
CatchAndRethrowDeep | Hardware | 1,092.07 us | 1,083.54 us | -0.78%
CatchAndThrowOtherDeep | Hardware | 1,192.87 us | 1,168.42 us | -2.05%
TryAndFinallyDeep | Hardware | 345.16 us | 332.23 us | -3.75%
TryAndCatchDeep_CaugtAtTheTop | Hardware | 311.23 us | 298.14 us | -4.21%
ThrowAndCatch | ReflectionSoftware | 401.58 us | 393.72 us | -1.96%
ThrowAndCatch_ManyCatchBlocks | ReflectionSoftware | 408.22 us | 399.12 us | -2.23%
ThrowAndCatchDeep | ReflectionSoftware | 584.56 us | 569.26 us | -2.62%
ThrowAndCatchDeepRecursive | ReflectionSoftware | 597.26 us | 584.98 us | -2.06%
ThrowAndCatch | ReflectionHardware | 414.25 us | 408.86 us | -1.30%
ThrowAndCatch_ManyCatchBlocks | ReflectionHardware | 422.66 us | 412.65 us | -2.37%
ThrowAndCatchDeep | ReflectionHardware | 595.31 us | 580.93 us | -2.42%
ThrowAndCatchDeepRecursive | ReflectionHardware | 619.57 us | 600.88 us | -3.02%



Benchmark: https://github.com/dotnet/performance/blob/e8a8fc62be32783d62fabffb56d14874d92a87cc/src/benchmarks/micro/libraries/System.Text.Json/Utf8JsonWriter/Perf.Basic.cs

|           Method    |  Formatted  |  SkipValidation  |  DataSize  | Before           | After            | Change |
|---------------------|-------------|------------------|------------|------------------|------------------|--------|
|     WriteBasicUtf8  |      False  |           False  | 10         |        837.8 ns  |        823.7 ns  | -1.68% |
|  WriteBasicUtf16    |      False  |           False  | 10         |        929.4 ns  |        897.5 ns  | -3.43% |
|     WriteBasicUtf8  |      False  |           False  | 100000     |  1,861,030.1 ns  |  1,813,554.0 ns  | -2.55% |
|  WriteBasicUtf16    |      False  |           False  | 100000     |  1,809,140.0 ns  |  1,802,474.4 ns  | -0.37% |
|     WriteBasicUtf8  |      False  |            True  | 10         |        777.7 ns  |        783.3 ns  | 0.72%  |
|  WriteBasicUtf16    |      False  |            True  | 10         |        848.1 ns  |        852.4 ns  | 0.51%  |
|     WriteBasicUtf8  |      False  |            True  | 100000     |  1,685,964.1 ns  |  1,734,788.7 ns  | 2.90%  |
|  WriteBasicUtf16    |      False  |            True  | 100000     |  1,682,827.2 ns  |  1,675,294.4 ns  | -0.45% |
|     WriteBasicUtf8  |       True  |           False  | 10         |      1,022.4 ns  |      1,003.1 ns  | -1.89% |
|  WriteBasicUtf16    |       True  |           False  | 10         |      1,097.4 ns  |      1,079.5 ns  | -1.63% |
|     WriteBasicUtf8  |       True  |           False  | 100000     |  2,471,009.2 ns  |  2,494,962.1 ns  | 0.97%  |
|  WriteBasicUtf16    |       True  |           False  | 100000     |  2,480,929.4 ns  |  2,478,218.4 ns  | -0.11% |
|     WriteBasicUtf8  |       True  |            True  | 10         |        961.6 ns  |        970.3 ns  | 0.90%  |
|  WriteBasicUtf16    |       True  |            True  | 10         |      1,039.3 ns  |      1,035.0 ns  | -0.41% |
|     WriteBasicUtf8  |       True  |            True  | 100000     |  2,398,607.5 ns  |  2,403,773.9 ns  | 0.22%  |
|  WriteBasicUtf16    |       True  |            True  | 100000     |  2,379,204.7 ns  |  2,376,995.1 ns  | -0.09% |

Benchmark: https://github.com/dotnet/performance/blob/4e14ecc54759224121b807c70e46fb1d9ee0e7d6/src/benchmarks/micro/runtime/BenchmarksGame/fasta-1.cs

Method | Before | After | Change
-- | -- | -- | --
Fasta_1 | 646.9 us | 633.1 us | -2.18%